Revert "MapDef: Don't use OpenCL type in kernel definition"
This reverts commit aa1eac0030321f18f93f7709ba6fb70074e88043.

+++ include/SkelCL/detail/MapDef.h
@@ -203,7 +203,7 @@ __kernel void SCL_MAP(
     const unsigned int          SCL_ELEMENTS,
     const unsigned int          PARTITION_SIZE)
 {
-  unsigned int i = get_global_id(0);
+  size_t i = get_global_id(0);
 
   while (i < SCL_ELEMENTS) {
     SCL_OUT[i] = SCL_FUNC(SCL_IN[i]);
